[Journal] Feedback updates
