+++ types/moderation.ts
@@ -1,7 +1,7 @@
 import { createTransformer } from './transformer';
 import { transformAuthorProfile } from '@/types/authorProfile';
 import { transformApplication } from '@/types/funding';
-import { FeedEntry, FeedGrantContent, RawApiFeedEntry } from '@/types/feed';
+import { FeedEntry, FeedGrantContent, RawApiFeedEntry, transformFeedEntry } from '@/types/feed';
 import type { GrantStatus } from '@/types/grant';
 import { transformTopic } from '@/types/topic';
 import { stripHtml } from '@/utils/stringUtils';
@@ -137,6 +137,24 @@ export const transformPendingGrantToFeedEntry = (entry: RawApiFeedEntry): FeedEn
     hotScoreBreakdown: entry.hot_score_breakdown,
   };
 };
+
+export function transformRegisteredReportCandidate(entry: RawApiFeedEntry): FeedEntry | null {
+  try {
+    const timestamp =
+      entry.action_date || entry.created_date || entry.content_object?.created_date || '';
+    const candidate = transformFeedEntry({
+      ...entry,
+      action: entry.action || 'PUBLISH',
+      action_date: timestamp,
+      created_date: entry.created_date || timestamp,
+      recommendation_id: entry.recommendation_id ?? null,
+    });
+
+    return candidate.content.contentType === 'PREREGISTRATION' ? candidate : null;
+  } catch {
+    return null;
+  }
+}
 
 // Hook state types
 export interface UserModerationState {
