+++ types/note.ts
@@ -7,7 +7,6 @@ import { Fundraise, transformFundraise } from './funding';
 import { Topic, transformTopic } from './topic';
 import { Grant, transformGrant } from './grant';
 import { AuthorProfile, transformAuthorProfile } from './authorProfile';
-import { normalizeRegisteredReportId } from '@/utils/registeredReportPrefill';
 export type NoteAccess = 'WORKSPACE' | 'PRIVATE' | 'SHARED';
 
 export type Author = {
@@ -120,9 +119,7 @@ export const transformNote = createTransformer<any, Note>((raw) => ({
   isRemoved: raw.unifiedDocument?.isRemoved || false,
   post: raw.post ? transformPost(raw.post) : null,
   documentType: raw.document_type ?? null,
-  proposalId:
-    normalizeRegisteredReportId(raw.proposal_id) ??
-    normalizeRegisteredReportId(raw.registered_report_prefill?.proposal_id),
+  proposalId: raw.proposal_id ?? raw.registered_report_prefill?.proposal_id ?? null,
 }));
 
 export const transformNoteWithContent = createTransformer<any, NoteWithContent>((raw) => ({
