--- utils/registeredReportPrefill.ts
@@ -1,9 +1,0 @@
-export function normalizeRegisteredReportId(value: unknown): number | null {
-  if (typeof value === 'number') {
-    return Number.isSafeInteger(value) && value > 0 ? value : null;
-  }
-  if (typeof value !== 'string' || !/^\d+$/.test(value)) return null;
-
-  const normalized = Number(value);
-  return Number.isSafeInteger(normalized) && normalized > 0 ? normalized : null;
-}
